+++ src/client/Client.ts
@@ -126,7 +126,7 @@ class WrapperClient extends CustomEvent {
         //config
         this.config = new Object({ ..._defaultConfig, ...config });
 
-        if (this.config.region) {
+        if (config.region) {
             this.lockRegion = true;
         } else {
             this.lockRegion = false;
@@ -134,6 +134,8 @@ class WrapperClient extends CustomEvent {
 
         if (this.config.region === 'data') {
             this.emit('error', { errorCode: 'ValWrapper_Config_Error', message: 'Region Not Found', data: this.config.region });
+            this.config.region = 'na';
+        } else if (!this.config.region) {
             this.config.region = 'na';
         }
 
@@ -146,12 +148,16 @@ class WrapperClient extends CustomEvent {
         this.entitlements_token = '';
         this.region = {
             pbe: 'na',
-            live: String(this.config.region),
+            live: this.config.region,
         };
         this.multifactor = false;
         this.isError = false;
 
         // first reload
+        if(this.lockRegion === false && this.config.region){
+            this.region.live = this.config.region;
+        }
+
         this.RegionServices = new WrapperRegion(this.region.live as keyof typeof _Region).toJSON();
 
         const _axiosConfig: AxiosRequestConfig = {
@@ -190,6 +196,10 @@ class WrapperClient extends CustomEvent {
      * @returns {void}
      */
     private reload(): void {
+        if(this.lockRegion === false && this.config.region){
+            this.region.live = this.config.region;
+        }
+            
         this.RegionServices = new WrapperRegion(this.region.live as keyof typeof _Region).toJSON();
 
         const _axiosConfig: AxiosRequestConfig = {
@@ -259,10 +269,6 @@ class WrapperClient extends CustomEvent {
         this.entitlements_token = data.entitlements_token;
         this.region = data.region;
 
-        if (this.region.live) {
-            this.lockRegion = true;
-        }
-
         this.reload();
     }
 
@@ -295,14 +301,17 @@ class WrapperClient extends CustomEvent {
         this.cookie = CookieJar.fromJSON(JSON.stringify(auth.cookie));
         this.access_token = auth.access_token;
         this.id_token = auth.id_token;
-        this.expires_in = auth.expires_in;
+        this.expires_in = auth.expires_in || Number(3600);
         this.token_type = auth.token_type;
         this.entitlements_token = auth.entitlements_token;
-        if (!this.lockRegion) {
-            this.region = auth.region;
+        this.region = auth.region;
+
+        if(!this.region.live) {
+            this.region.live = 'na';
         }
-        this.multifactor = auth.multifactor;
-        this.isError = auth.isError;
+
+        this.multifactor = auth.multifactor || Boolean(false);
+        this.isError = auth.isError || Boolean(false);
 
         if (auth.isError && !this.config.forceAuth) {
             this.emit('error', {
@@ -349,6 +358,7 @@ class WrapperClient extends CustomEvent {
     public setRegion(region: keyof typeof _Region): void {
         this.emit('changeSettings', { name: 'region', data: region });
 
+        this.config.region = region;
         this.region.live = region;
         this.reload();
     }
@@ -395,7 +405,7 @@ class WrapperClient extends CustomEvent {
     //static
 
     /**
-     * 
+     * * Something went wrong? try to not use static methods.
      * @param {ValWrapperConfig} config Client Config
      * @param {ValWrapperClient} data Client `.toJSON()` data
      * @returns {WrapperClient}
@@ -404,15 +414,18 @@ class WrapperClient extends CustomEvent {
         const NewClient: WrapperClient = new WrapperClient(config);
         NewClient.fromJSON(data);
 
-        NewClient.expires_in = 3600;
-        NewClient.multifactor = false;
-        NewClient.isError = false;
+        if(config.region) {
+            NewClient.setRegion(config.region);
+        } else if (data.region.live) {
+            NewClient.setRegion(data.region.live as keyof typeof _Region);
+        }
 
         return NewClient;
     }
 
     /**
-     * * Not Recommend
+     * * Not Recommend to use
+     * * After run this method, you must use `.setRegion()` to set region.
      * @param {ValWrapperConfig} config Client Config
      * @param {ValWrapperAuth} data Authentication Data
      * @returns {Promise<WrapperClient>}
@@ -430,9 +443,13 @@ class WrapperClient extends CustomEvent {
             isError: data.isError,
         }
 
-        const NewCookieAuth = await ClientAuthCookie.reauth(CookieAuthData, String(config.userAgent), _Client_Version, toUft8(JSON.stringify(_Client_Platfrom)));
+        const NewCookieAuth = await ClientAuthCookie.reauth(CookieAuthData, config.userAgent || String(_defaultConfig.userAgent), _Client_Version, toUft8(JSON.stringify(_Client_Platfrom)));
 
-        return WrapperClient.fromJSON(config, NewCookieAuth);
+        //client
+        const NewClient: WrapperClient = new WrapperClient(config);
+        NewClient.fromJSONAuth(NewCookieAuth);
+
+        return NewClient;
     }
 }